Provide a simple prefetcher implementation.
This functionality is available in `SDWebImage` which influenced this
library, and was something we needed in migrating over.

It is a simple implementation, using default managers etc. but could be
expanded upon pretty easily to use different managers, downloaders, and
queues etc.

A utility method to check for cached images (useful when deciding
whether to prefetch or not) has also been added to `KingfisherManager`.

All new code is tested.