Adds support to download files via NSURLSession.downloadTask.
Adds a .useFileDownloads to  KingfisherOptionsInfo.   The feature is only enabled when asking for it.

Files are then directly MOVED In to the file cache.
Is a faster version of 'cacheOriginalDownload'.   This may help dealing with downloading images that might be TOO BIG to store in RAM.

Files are directly moved/copied into their target imageCaches file space directly.